+++ node/resolvers/index.ts
@@ -1,7 +1,11 @@
 import { v4 as uuidv4 } from 'uuid'
+import { JSDOM } from 'jsdom'
+import createDOMPurify from 'dompurify'
+// eslint-disable-next-line prettier/prettier 
+import type { ServiceContext } from '@vtex/api'
 
 import FileManager from '../FileManager'
-import { ServiceContext } from '@vtex/api'
+
 
 type FileManagerArgs = {
   path: string
@@ -17,36 +21,49 @@ type UploadFileArgs = {
 }
 
 const isValidFileFormat = (extension: string, mimetype: string) => {
-  const allowedExtensions = [
-    'png',
-    'jpg',
-    'jpeg',
-    'gif',
-    'webp',
-    'pdf',
-    'doc',
-    'docx',
-    'svg',
-    'xls',
-    'xlsx',
-    'txt'
-  ]
-  
-  const allowedMimeTypes = [
-    'image/png',
-    'image/jpeg',
-    'image/gif',
-    'image/webp',
-    'image/svg+xml',
-    'application/pdf',
-    'application/msword',
-    'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
-    'application/vnd.ms-excel',
-    'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
-    'text/plain'
-  ]
 
-  return allowedExtensions.includes(extension) && allowedMimeTypes.includes(mimetype)
+  if (!extension || !mimetype) {
+    return false
+  }
+
+  // Normalize extension to lowercase
+  const normalizedExtension = extension.toLowerCase()
+
+  // Define allowed file types with their corresponding MIME types
+  const allowedFileTypes = {
+  png: 'image/png',
+  jpg: 'image/jpeg',
+  jpeg: 'image/jpeg',
+  gif: 'image/gif',
+  webp: 'image/webp',
+  svg: 'image/svg+xml',
+  pdf: 'application/pdf',
+  doc: 'application/msword',
+  docx: 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
+  xls: 'application/vnd.ms-excel',
+  xlsx: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
+  txt: 'text/plain',
+}
+
+  if(!(extension in allowedFileTypes)) {
+    return false
+  }
+
+  return normalizedExtension in allowedFileTypes && allowedFileTypes[normalizedExtension as keyof typeof allowedFileTypes] === mimetype
+}
+
+const isValidSVGFile = async (loadedFile: any) => {
+    const fileBuffer = await loadedFile.createReadStream().toArray()
+    const fileString = Buffer.concat(fileBuffer).toString('utf8')
+        
+    const {window} = new JSDOM('')
+    const DOMPurify = createDOMPurify(window)
+
+    const cleanSvgString = DOMPurify.sanitize(fileString, {
+      USE_PROFILES: { svg: true },
+    })
+        
+    return cleanSvgString === fileString
 }
 
 export const resolvers = {
@@ -55,7 +72,8 @@ export const resolvers = {
       const fileManager = new FileManager(ctx.vtex)
       const { path, width, height, aspect, bucket } = args
 
-      const file = await fileManager.getFile(path, width, height, aspect, bucket)
+      const file = await fileManager.getFile({path, width, height, aspect, bucket})
+
       return file
     },
     getFileUrl: async (_: unknown, args: FileManagerArgs, ctx: ServiceContext) => {
@@ -63,6 +81,7 @@ export const resolvers = {
       const { path, bucket } = args
 
       const file = await fileManager.getFileUrl(path, bucket)
+
       return file
     },
     settings: async () => ({
@@ -79,6 +98,18 @@ export const resolvers = {
 
       if (!isValidFileFormat(extension, mimetype)) {
         throw new Error('Invalid file format') 
+      }
+
+      // Validate SVG files separately
+      // SVG files require additional validation to prevent XSS attacks
+      // and other security issues, so we check if the file is SVG
+      // and sanitize it if necessary.         
+
+      if (mimetype === 'image/svg+xml') {     
+        const isSafe = await isValidSVGFile(loadedFile)
+          if (!isSafe) {            
+            throw new Error('Invalid or malicious SVG file')
+          }   
       }
 
       const filename = `${uuidv4()}.${extension}`